refactor to one validating parent class
